refactor(client): isolate code dedicated to running tests
